[refactor]: controllers dir의 파일 명 수정 및 코드 리팩터 #2
