+++ server/src/routes/auth.js
@@ -2,12 +2,12 @@ const express = require('express');
 const passport = require('passport');
 
 const router = express.Router();
-const { githubAuth } = require('../controllers/github');
-const { revokeAuth } = require('../controllers/logout');
+const AuthController = require('../controllers/auth');
 
 router.get('/github/login', passport.authenticate('github'));
-router.get('/github/callback', githubAuth);
+router.get('/github/callback', AuthController.githubAuth);
 
-router.get('/logout', revokeAuth);
+// TODO: logout을 위한 controller 작성 필요
+// router.get('/logout',);
 
 module.exports = router;
